+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Require php 8.1
+- `Innmind\SshKeyProvider\Local` now uses the filesystem abstraction instead of using processes
 
 ### Removed